% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambia_forma_eta.R
\name{cambia_forma_eta}
\alias{cambia_forma_eta}
\title{Prende come argomento un vettore di classi di età nel formato "(14,19]" e lo
restituisce in formato "15-19" (se right = TRUE). Se right = FALSE, prende
"[15,20)" e restituisce "15-19"}
\usage{
cambia_forma_eta(classi_eta, right = FALSE)
}
\arguments{
\item{classi_eta}{vettore (character o factor) tipo "(aa,aa]" (o "[aa,aa)") che indica gli estremi della classe di età in esame}

\item{right}{logical (TRUE o FALSE), che indica se l'intervallo comprende l'estremo superiore (default = FALSE, quindi che l'intervallo comprende il limite inferiore)}
}
\description{
Prende come argomento un vettore di classi di età nel formato "(14,19]" e lo
restituisce in formato "15-19" (se right = TRUE). Se right = FALSE, prende
"[15,20)" e restituisce "15-19"
}
\examples{
cambia_forma_eta("[15,20)")
cambia_forma_eta("(14,19]", right = TRUE)
}
