% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea_coorti.R
\name{crea_coorti}
\alias{crea_coorti}
\title{Trasforma un vettore numerico (possibilmente di numeri interi) in un factor (classi).}
\usage{
crea_coorti(x, breaks, include.lowest = TRUE, right = FALSE, ...)
}
\arguments{
\item{x}{un vettore numerico (da cui si desiderano delle classi)}

\item{breaks}{either a numeric vector of two or more unique cut points or a single number (greater than or equal to 2) giving the number of intervals into which x is to be cut.}

\item{include.lowest}{logical, indicating if an 'x[i]' equal to the lowest (or highest, for right = FALSE) 'breaks' value should be included.}

\item{right}{logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Come \code{base::cut} con i parametri \code{include.lowest = TRUE} e \code{right = FALSE} (di default)
}
\examples{
crea_coorti(x = seq(1985, 2000, by = 1), breaks = c(1985, 1990, 1995, 2000, Inf))
}
\seealso{
\code{cut}
}

