% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta2coorti.R
\name{eta2coorti}
\alias{eta2coorti}
\title{Trasforma una classe di et\`a in una coorte}
\usage{
eta2coorti(classe_eta, anno, right = FALSE)
}
\arguments{
\item{classe_eta}{vettore (character o factor) con la classe d'et\`a ("[15,20)" o "(15,20]")}

\item{anno}{anno di riferimento, che serve per dedurre l'anno di nascita dall'et\`a in anni}

\item{right}{logical (TRUE o FALSE). Indica se la classe di et\`a comprende l'estremo superior (right = TRUE) oppure no (right = FALSE, opzione di defauls)}
}
\description{
Trasforma una classe di et\`a in una coorte
}
\examples{
eta2coorti("[15,20)", anno = 2000)
eta2coorti("[15,20)", anno = 2000, right = TRUE)

}

