% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_and_compress.R
\name{export}
\alias{export}
\title{Esporta un oggetto (di solito un data frame) in un file di testo.}
\usage{
export(.data, name = NULL, fileEncoding = "CP1252", ...)
}
\arguments{
\item{.data}{oggetto da esportare (data frame)}

\item{name}{nome del file in cui esportare (default NULL, crea un nome automaticamente sulla base della data)}

\item{fileEncoding}{encoding del file (default "CP1252", per windows...). Altro valore utilizzato spesso: "UTF-8".}

\item{...}{altri argomenti che si possono passare (vedi \code{utils::write.table})}
}
\description{
Con ";" come separatore di colonne.
}
