% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punto_migliaia.R
\name{first_m}
\alias{first_m}
\title{Funzione che cambia in maiuscolo la prima lettera di una stringa di testo}
\usage{
first_m(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
Funzione che cambia in maiuscolo la prima lettera di una stringa di testo
}
\examples{
first_m("pippo")

}

