% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibr-package.R
\docType{package}
\name{gibr-package}
\alias{gibr}
\alias{gibr-package}
\title{gibr: Funzioni Varie (Senza uno Scopo Particolare)}
\description{
Funzioni varie (senza uno scopo particolare). Un modo per salvare varie funzioni create per poi averle a disposizione attraverso un pacchetto.
}
\keyword{internal}
