% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punto_migliaia.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Operatore \%nin\% (l'opposto di \%in\%)}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vettore da mettere a sinistra.}

\item{y}{vettore da mettere a destra (si valuta se gli elementi di questo vettore non fanno parte del vettore x).}
}
\description{
Operatore \%nin\% (l'opposto di \%in\%)
}
\examples{
1:10 \%in\% c(1,3,5,9)
1:10 \%nin\% c(1,3,5,9)
}
