% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punto_migliaia.R
\name{punto_migliaia}
\alias{punto_migliaia}
\title{Format numbers with . as thousand separator and comma for decimal separator}
\usage{
punto_migliaia(x, nsmall = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{nsmall}{(as in the \code{base::format} function)}
}
\value{
A character vector with the numbers formatted
}
\description{
Format numbers with . as thousand separator and comma for decimal separator
}
\examples{
x <- runif(10) * 5000
punto_migliaia(x)

}
