% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_templates.R
\name{rmd_yaml}
\alias{rmd_yaml}
\alias{rmd_yaml_beamer}
\title{Crea un esempio di codice 'yaml' per iniziare un documento ".Rmd"}
\usage{
rmd_yaml()

rmd_yaml_beamer()
}
\value{
un vettore character che si può esportare in un file con \code{link{writeLines}}
}
\description{
Questa funzione non ha parametri. Genera solo un esempio di codice 'yaml'
da includere all'inizio di un documento '.Rmd', con alcune opzioni che usiamo
ogni tanto.
}
\examples{
doc <- rmd_yaml()

doc

# Esporazione in un file
writeLines(doc, con = "documento.Rmd")
}
