% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punto_migliaia.R
\name{round2}
\alias{round2}
\title{Funzione equvalente a round, tranne per l'arrotondamento di .5 all'intero più grande (round invece, ad esempio, arrotonda 2.5 a 2)
Trovata su stackoverflow
http://stackoverflow.com/questions/12688717/round-up-from-5-in-r}
\usage{
round2(x, n)
}
\arguments{
\item{x}{un vettore numerico}

\item{n}{il numero di decimali per cui si vuole arrotondare}
}
\description{
Funzione equvalente a round, tranne per l'arrotondamento di .5 all'intero più grande (round invece, ad esempio, arrotonda 2.5 a 2)
Trovata su stackoverflow
http://stackoverflow.com/questions/12688717/round-up-from-5-in-r
}

